pg_upgrade: drop all tables that cause relation mismatches
A continuation of the previous commit -- drop the rest of the tables
that cause pg_upgrade to complain about relation mismatches after
performing an upgrade. With this commit, the 6->6 upgrade smoke test
goes green on master, and we can start working on adding these dropped
tables back one by one.