Completely remove interconnect kernel module related test case
We used to use kernel module named ickm to do functional test of
interconnect, an example is: ickm capature a UDP packet in kernel,
if this packet is a data packet, drop it, then ickm will capture
another packet and will test that this packet is a resent packet.

Those kernel module related cases have many defects:

it relies on tinc to do kernel module installation/deletion
it needs root permission
it's kernel version sensitive to compile the module
the case itself is not stable
In the current pipeline, those jobs are actually been skipped due
to platform mismatch, but it still make troubles when we want to
integrate the interconnect job to another platform like Oracle7,
so this commit will remove those kernel module related tests completely.

Next step:
We need to do similar functional test for interconnect but instead
of kernel module, to make it more dependent and stable.