Make use of serial/deserial functions to enable 2-phase aggregates.
Previously, two-phase (or three-phase) aggregation was not possible for
aggregates with 'internal' transition type. In GPDB5, some aggregates
had been modified to use bytea as the transition type, to work around that
problem, but now we can use the serial/deserial functions like in the
upstream.

This commit doesn't include any serial/deserial functions yet. They will
be introduced in subsequent commits.