Fix CREATE OR REPLACE LANGUAGE.
The 'replace' field was missing from the out/read functions. Because of
that, if the language already existed, the command would fail on the
segments with "language already exists" error.

For the sake of completeness, add a test, although it seems pretty unlikely
for this particular bug to reappear in the future.

No backpatching required, as the OR REPLACE option was added in PostgreSQL
9.0.