Enable mapred debugging and remove GUC in core
The gp_mapred_define GUC was intended to control statement logging
during gpmapreduce debugging, but it could only be twiddled in the
gpmapreduce code with a specialized debug build. This hardly seems
worth to keep an extension GUC in core for, with tests for it in
reasonably hot codepaths, so remove it.

While at it, enable the debugging tools in gpmapreduce from normal
builds since they are harmless and could potentially be useful.

Reviewed-by: Asim R P <apraveen@pivotal.io>