Support PREFUNC as an alias for COMBINEFUNC, for backwards compatibility.
Commit b8545d57a5 renamed CREATE AGGREGATE's PREFUNC option into
COMBINEFUNC, because that's what it's called in the upstream. However,
there might be old SQL scripts lying around that still the old PREFUNC
option. CREATE AGGREGATE gives only a warning about unrecognized options,
so such a command would still succeed, and create the aggregate, but
without the combine function. As a courtesy to users upgrading from GPDB 5
and below, support PREFUNC as an alias for COMBINEFUNC.

In the passing, remove PREFUNC from the \h page for CREATE AGGREGATE. That
was meant to be removed in commit b8545d57a5, and I don't think should list
backwards-compatibility options like this in \h. Add a note to the dita docs
for it, instead.

Discussion: https://groups.google.com/a/greenplum.org/d/msg/gpdb-dev/Aq3aeA3Jm3w/sTeQs2iMBAAJ