Remove unnecessary gpaotid datatype.
'gpaotid' was similar to 'tid', but for AO tables. However, it was not
used anywhere. You could cast a value from 'tid' to 'gpaotid', and the
output function would then print the value in (segment #, row #) format,
but that's not very useful. And it was not documented.

If we need something like that for debugging purposes, a function to
"decode" an AO tid value from the 'tid' format to AO segment # and row #
might be useful. We have an internal function AOTupleIdToString() to do
that, which is used in some debugging messages, so perhaps it would make
sense to expose that at the SQL level. But a whole separate datatype seems
overblown, and I don't see much need for even a function at the moment,
so let's just rip this out.