Re-enable the CDB planner for DISTINCT-qualified aggregates.
In the Window functions rewrite, I had accidentally disabled the creation
of the multi-phase DQA plans. As a result, queries containing DISTINCT-
qualified aggregates, e.g. "SUM(DISTINCT n)", generated naive single-phase
plans, which can be much less efficient.

In PostgreSQL, DISTINCT aggregates are planned similarly to ordered
aggregates, using a tuplesort inside the Agg node. Because of that, the
AggClauseCosts.numOrderedAggs count includes both DISTINCT and ordered
aggregates. With cdb_grouping_planner, we fall back to that method for
ordered aggregates, but for DISTINCT aggregates, it can generate smarter
plans. However, because it was checking numOrderedAggs to check whether
the plan contained any, it got fooled and also fell back if there were any
DISTINCT aggregates. To fix, add a separate counter for true ordered
aggregates that doesn't include DISTINCTs.

The DQA-planner code had bitrotted somewhat, while it's not been used.
Nowadays, the Aggref.args list contains TargetEntrys instead of plain
expressions. And when constructing subqueries on the fly, the RelOptInfos
created for the RTE_SUBQUERY rtable entries need to have the 'subplan'
and 'subroot' fields filled in. Those changed were made to the similar code
for the normal, non-DISTINCT 2-phase aggs during the merges, but the
DQA code was neglected, because it was unused.

We have tests for DQAs, but apparently none of the tests use EXPLAIN, so
although you still got correct results, we did not notice that the plans
regressed. The expected output of a few existing EXPLAIN tests changes but
they were fairly simple cases where the naive plan actually seems fine as
well. Add EXPLAINs to a few queries in the gp_dqa test, so that we will be
alarmed if this happens again.

Fixes https://github.com/greenplum-db/gpdb/issues/5627