Support FORCE QUOTE * in external tables
FORCE QUOTE * was added in PostgreSQL 9.0 as a shorthand for FORCE
QUOTE <column_1 .. column_n> in the COPY comamnd, where all columns
of a relation are added for forced quoting. External tables use copy
and copy options under the hood, so they too should support '*' to
quote all columns. This resolves a FIXME added during the 9.0 merge.