Move qp_functions_in_* tests to another parallel group.
Much of the time in qp_functions_in_* tests is spent on waiting, after an
error occurred. There's a one second delay after an error happens in a
segment, until the dispatcher reacts to it. We can therefore reduce the
overall runtime of the regression suite by running them in parallel with
other tests that take an even longer time, hiding the waits behind the
other tests that do real work. This should shave off 10-20 seconds from
the overall regression test suite time. Every little helps.