pg_upgrade: diff schema during smoke test
We don't upgrade the segments during the ICW smoke test, but we can
still connect to the master and dump the schema to make sure
everything's okay. This will let us put some basic pg_dump regression
cases into ICW.

A few new failures fall out as a result. In keeping with the desire to
parallelize fixes to pg_dump/_upgrade, I'm dropping them before the
upgrade tests and we can add them back as they are fixed.

Also remove the post-test cleanup step, as it's not currently being
used.