pg_upgrade: Fix dump file diff caused by default value with type bit varying in column of a relation. (#4823)
Here is the repro case:
psql -X -d regression -c "CREATE TABLE t111 ( a40 bit varying(5) DEFAULT '1');"
psql -X -d regression -c "CREATE TABLE t222 ( a40 bit varying(5) DEFAULT B'1');"

After pg_upgrade testing, we will see failure and the diff of dump sql files looks like this:

 CREATE TABLE t111 (
-    a40 bit varying(5) DEFAULT B'1'::bit varying
+    a40 bit varying(5) DEFAULT (B'1'::"bit")::bit varying
 ) DISTRIBUTED BY (a40);

No diff for the table t222. From perspective of functionality, the difference
seems to mean nothing, but it is annoying for CI.

This issue was found when testing pg_upgrade using the pg_regression database which is generated after running regression tests so we do not need to add a test case for it.

This issue exists on latest PG also. We submitted a patch to upstream and the patch is under reviewing. We'll check in GP in advance.

Co-authored-by: Richard Guo <riguo@pivotal.io>
Co-authored-by: Paul Guo <pguo@pivotal.io>