Correct cost calculation for SplitUpdate plan (#5698)
Correct cost calculation for SplitUpdate plan

This partly addresses a GPDB_90_MERGE_FIXME introduced in 73801e8. As mentioned
in the FIXME, this will not help generating a better plan because we have no
choice other than simply adding the SplitUpdate node. Note that only the cost
is adjusted, the width is still incorrect. We will not fix width for now
because upstream commit 3fc6e2d will fix it.

Co-authored-by: Shujie Zhang <shzhang@pivotal.io>
Co-authored-by: Alexandra Wang <leiwangcheme@gmail.com>