Small cleanups of with_clause test suite
Use EXPLAIN output to test the qual pushdown rather than just invoking
the query, as it's hard to varify the pushdown from examining that (this
introduces an _optimizer file). Also remove redundant DROP clauses,
refactor a few things to avoid NOTICEs and extend documentation to not
just list an internal Jira ticket.