Fix ao_upgrade tests and add them to the schedule
9.1 added a new, more compact "short" format to the numeric datatype.
This format wasn't handled by the ao_upgrade test in isolation2, so it
failed -- but the pipeline was still green because I forgot to add the
new test to the schedule in 54895f54.

To fix the issue, add a new helper which will force any Numeric back to
the legacy long format, and call that from convertNumericToGPDB4() in
the ao_upgrade test. And add the test to the schedule, so we don't have
to do this again.

Co-authored-by: Kalen Krempely <kkrempely@pivotal.io>