Update gpsd and minirepro to dump HLL stats as an option
Currently `gpsd` and `minirepro` dumps the hll stats aaand it causes the
output files to have bigger size. As we use these tools to debug
query plans, we do not use the HLL counter info in the query planning.
Instead it is just used to derive root level stats for partition tables.
For this reason, it is better to provide HLL stats dump as an option
instead of having `gpsd` and `minirepro` dump it by default.

This commit addresses this issue.