Move zstd ICG test into its gpcontrib module
The test is included in "installcheck" under the zstd module. It should
be eventually included as part of ICW.

CAVEAT: the test for zstd, as it turns out, is wrong (nondeterministic)
since its inclusion in commit 724f9d27b16.

This move eliminiates the need to have a separate "error: zstd not
supported" answer file in ICG.

Co-authored-by: Jesse Zhang <sbjesse@gmail.com>