Remove mock test for cdbappendonlystorage.c
Looking at the test, it seems pretty uninteresting. It basically tests that
there is call to IsStandbyMode() in appendonly_redo(), and that the
switch-case statement works. I don't think we need bespoken tests for those
things.

This came up while working on the 9.3 merge. 9.3 moves all the *_desc()
functions to separate files, so after that, only the appendonly_redo()
function would remain in cdbappendonlystorage.c. It would make sense to
move it into cdbappendonlyxlog.c, but we couldn't do that, because this
mock test got in the way. We could fix that in the merge, but I'm opening
a separate PR of this removal, to give more visibility to the decision.