Add error position to error on coercing partition boundaries.
I think this is more clear. Now that we the error position is displayed
with the error, we don't need to work so hard to deparse the value and the
partition type. The error position makes that clear.

This came up during the 9.3 merge. Passing InvalidOid to
deparse_context_for() will no longer work in 9.3, so we'd need to change
this somehow in the 9.3 merge, anyway.