Fix pgbench.
It was left broken by the 9.2 merge. Need to use getopt_long() now, as in
upstream, and we were passing wrong variable to snprintf(). Both of these
were causing compiler warnings, and the latter was causing a segfault.

We don't seem to have any tests for pgbench. I think we'll get some from
upstream later on, and pgbench isn't super-critical anyway, so I didn't
address that at this moment.