Remove the printing of overflowed value, when scale is exceeded.
Printing the value was added in GPDB, back in 2007. The commit message of
that change (in the historical pre-open-sourcing git repository) said:

    Merge forward from Release-3_0_0-branch. Update comment block.
    Tidy numeric_to_pos_int8_trunc.

That wasn't not very helpful...

Arguably, printing the value can be useful, but if so, we should submit
this change to upstream. I don't think it's worth the trouble, though, so
I suggest that we just revert this to the way it is in the upstream. The
reason I'm doing this now is that this caused merge conflicts in the 9.3
merge, that we're working on right now. We could probably fix the conflict
in a way that keeps the extra message, but it's simpler to just drop it.