Fix pg_dump and pg_dumpall issues relating to gp_default_storage_options
When pg_dump or pg_dumpall dumps a database that has gp_default_storage_options
GUC set to something other than the default settings, restoring the tables in
that database immediately after connecting to it would possibly create the table
with the wrong storage parameters. To prevent this, we make sure the
gp_default_storage_options GUC is set to built-in default values before
restoring tables.

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>