Remove duplicated code to handle SeqScan, AppendOnlyScan and AOCSScan.
They were all treated the same, with the SeqScan code being duplicated
for AppendOnlyScans and AOCSScans. That is a merge hazard: if some code
is changed for SeqScans, we would have to remember to manually update
the other copies. Small differences in the code had already crept up,
although given that everything worked, I guess it had no effect. Or
only had a small effect on the computed costs.

To avoid the duplication, use SeqScan for all of them. Also get rid of
TableScan as a separate node type, and have ORCA translator also create
SeqScans.

The executor for SeqScan node can handle heap, AO and AOCS tables, because
we're not actually using the upstream SeqScan code for it. We're using the
GPDB code in nodeTableScan.c, and a TableScanState, rather than
SeqScanState, as the executor node. That's how it worked before this patch
already, what this patch changes is that we now use SeqScan *before* the
executor phase, instead of SeqScan/AppendOnlyScan/AOCSScan/TableScan.

To avoid having to change all the expected outputs for tests that use
EXPLAIN, add code to still print the SeqScan as "Seq Scan", "Table Scan",
"Append-only Scan" or "Append-only Columnar Scan", depending on whether
the plan was generated by ORCA, and what kind of a table it is.