Remove FIXME in RemoveLocalLock, it's alright.
The FIXME was added to GPDB in commit f86622d95a, which backported the
local cache of resource owners attached to LOCALLOCK. I think the comment
was added, because in the upstream commit that added the cache, the
upstream didn't thave the check guarding the pfree() yet. It was added
later in upstream, too, in commit 7e6e3bdd3c, and that had already been
backported to GPDB. So it's alright, the guard on the pfree is a good thing
to have, and there's nothing further to do here.