Delete SIGUSR2 based fault injection logic in walreceiver.
Regular fault injection doesn't work for mirrors. Hence, using SIGUSR2 signal
and on-disk file coupled with it just for testing a fault injection mechanism
was coded. This seems very hacky and intrusive, hence plan is to get rid of the
same. Most of the tests using this framework are found not useful as majority of
code is upstream. Even if needs testing, better alternative would be explored.