Refactor code in ProcessRepliesIfAny() to match upstream.
The reason we needed the FIXME pq_getmessage() call, marked with the
FIXME comment, was that we were missing the pq_getmessage() call from
ProcessStandbyMessage(), that the corresponding upstream version, at the
point that we're caught up in the merge, had. I believe the reason it was
missing from ProcessStandbyMessage() was that we had earlier backported
upstream commit cd19848bd55. That commit removed the pq_getmessage() call
from ProcessStandbyMessage(), and added one in ProcessRepliesIfAny(),
instead.

Clarify this by changing the code to match upstream commit cd19848bd55.
(Except that we don't have pq_startmsgread() yet, that will arrive when
we merge the rest of commit cd19848bd55.)