Fix UPDATE RETURNING on distribution key columns.
Updating a distribution key column is performed as a "split update", i.e.
separate DELETE and INSERT operations, which may happen on different nodes.
In case of RETURNING, the DELETE operation was also returning a row, and it
was also incorrectly counted in the row count returned to the client, in
the command tag (e.g. "UPDATE 2"). Fix, and add a regression test.

Fixes https://github.com/greenplum-db/gpdb/issues/5839