Correctly account for OptimizerOutstandingMemoryBalance
Because the optimizer has its own memory management system, and does not
make use of AllocationSets, the only way we know how much memory the
optimizer is using is by intercepting the calls to malloc() and free().
When the GUC 'optimizer_use_gpdb_allocators' is set to 'true' Orca will
replace its native  alloc() and free() methods with Ext_OptimizerAlloc()
and Ext_OptimizerFree(). These calls will track the total memory usage
in the active 'Optimizer' memory account, and the total outstanding
memory between queries in OptimizerOutstandingMemoryBalance.

This is a problem when accounting for ORCA memory in the
'X_NestedExecutor' account, because unless you add the
OptimizerOutstandingMemoryBalance ORCA can free memory that was
allocated in a previous query and underflow the account.

Both in order to get an accurate idea of how much memory the optimizer
is using, and to prevent problems with the 'X_NestedExecutor' account,
it makes sense to track the ORCA's memory usage in a single account.
Therefore, only create one 'Optimizer' account per query, no matter how
many times we call it.

Co-authored-by: David Kimura <dkimura@pivotal.io>