Update list of transaction, one distributed log page at a time.
This page addresses GPDB_84_MERGE_FIXME about updating the distributed log page
at a time. This makes it similar to how clog in updated efficiently in commit
06da3c570.

For committing the subtransactions in clog log more dance needs to be performed,
if all don't fall on same page. That kind of dancing is not required for
distributed log as just need to make sure all distributed log entries are made
before clog. All of them don't have to be atomic with respect to each other. So,
its very simple for distributed log.