Remove bkuprestore test
The bkuprestore test was imported along with the source code during the
initial open sourcing, but has never been used and hasn't worked in a
long time. Rather than trying to save this broken mess, let's remove it
and start fresh with a pg_dump TAP test which is a much better way to
test backup/restore.

Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>
Reviewed-by: Jimmy Yih <jyih@pivotal.io>