Complete backport of upstream commits around lastReplayedEdnRecPtr.
We had backported upstream commits 425bef6ee7 and 2cd72ba42d earlier, but
those got partially reverted in the 9.3 merge. Or earlier, or we hadn't
backported them completely to begin with - I didn't investigate the exact
path of how we got here. In any case, a partial backport is confusing, so
take the code around this from the tip of 9.3 stable, so that we have both
of those commits fully backported.