Refactor qp_functions test suite
The test suite, which was ported over from TINC, was ignoring so much of
the memorized output that it more or less didn't test anything (and the
ignored blocks was as full of outdated output as one would imagine). The
code was also formatted in weird ways and had needless NOTICEs thrown
during execution.

This refactors the testsuite to remove all ignore blocks, removes some
utterly pointless tests (there are many more of them left), formats the
code to be readable, fixes the output to work and removes some duplicate
tests.

The remaining bits of the suite is by no means terribly interestering,
but it runs fast enough that it's worth keeping the leftovers for now.

Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>