Fix the one-element tuple visibility cache in heap scans, for multi-xids.
It's not cool to use the raw xmax value as part of the cache key. If the
raw xmax represents a multi-xid, the real deleter XID would be something
else. We could get fooled, if we cached a multi-XID value, and later saw
a tuple with a regular xmax, with the same numerical value as the cached
multi-XID.

I think this was actually broken before the 9.3 merge already. If a
transaction locked a tuple, and deleted another tuple, and a concurrent
scan sees the locked tuple first, it might think that the deleted tuple
is also visible to it, because it has the same xmin+xmax combination as
the locked tuple.