Fix mixup between bitmap index's int48 and int42 support functions.
These were swapped. It's been wrong ever since we merged the operator
family patch, during the 8.3 merge. But apparently it wasn't causing any
ill effect, or at least I was not able to find a case that would fail
because of it.

This was caught by new sanity checks in the 'opr_sanity' regression
test, introduced in the upcoming 9.4 merge.