Remove unnecessary code for the first ORDER BY column in window agg.
The purpose of this code was to treat the first ORDER BY column, in a
window agg like "ROW_NUMBER() OVER (ORDER BY x RANGE BETWEEN 2 PRECEDING
AND 2 FOLLOWING", the same way as volatile expressions, and add them to
the target list as is. That was to ensure that it would be available for
computing the window bounds. But upstream commit a20993608a, merged as
part of the 9.3 merge, got rid of the distinction between volatile and
non-volatile expressions, so we no longer need to treat the first ORDER BY
column any different either.