Fix event triggers.
We punted on event triggers during the 9.3 merge. This patches up the
OID dispatch of various commands, to make event triggers work, and
re-enables the upstream test.

In GPDB, event triggers only fire in the QD node, not the segments. That
seems like the most sensible behavior. The pg_event_triggers catalog table
is not updated in segments, either, it will stay empty.

There's one remaining issue, marked with FIXME comments: The
pg_event_trigger_dropped_objects() function needs to run in the QD process,
or it errors out, but there's no way to force the planner to do that. So
if you try to use it as part of a more complicated query, it won't work.
One of the upstream tests had to be modified to work around that.