Also support event triggers on CREATE PROTOCOL.
We don't particularly care about event triggers on CREATE PROTOCOL as such,
but we were tripping the assertion in EventTriggerCommonSetup(), because
it was being called for CREATE PROTOCOL, even though it was listed as
not supported. It seems to actually work fine, if we just mark it as
suppported, so might as well.

Add a test case, also for the CREATE EXTERNAL TABLE support.