Guard against risk of string overflow in AO error codepath
The AoHeaderCheckErrorStr string is a statically allocated buffer,
so we must always use snprintf() when writing to it to reduce the
risk of an out-of-bounds write. This also moves the definition of
the error string to the top of the file to match project style.

Also remove unnecessary assertions on snprintf(); if we consider
it an error that message is truncated, then we should be using a
datastructure that supports being extended.

Reviewed-by: Jacob Champion <pchampion@pivotal.io>
Reviewed-by: Joao Pereira <jdealmeidapereira@pivotal.io>