Remove the ControlFileWatcher facility.
This was introduced in 3.x days for root causing a bug where writeLoc was less
than flushLoc. Didn't found this any helpful after that. Seems just an assertion
can cover the same is required instead of having this facility adding overhead
to work-load. Its not much but still writing to file, reading from file and
comparing. Hence, removing the same which helps further reduce diff compared to
upstream.