Check at planning time, whether an aggregate has serial/deserial functions.
If an aggregate, with transition type 'internal', is missing the serial or
deserial function, it cannot be used in partial aggregation. In principle,
it could still be done within a single process, but GPDB doesn't create
such plans, currently, so in practice the combine function is useless
without serial and deserial functions.

There is a check for this in the executor, but we better catch this at
planning time. All of the built-in aggregates have both combine and
serial/deserial functions, or neither, so this is only a problem for
user-defined aggregates.