Fail fast in replica check if primary and mirror disconnected.
If primary and mirror are not connected or not in STREAMING state, fail fast and
not retry. Retry only in-case they are connected but checkpoint is not streamed
across and yet to be replayed. Also, here reduce the waiting time from 1 sec
from checking to 1/5 of a sec so speed up.

Many times in CI replica check appeared to be hung because if primary and mirror
pair are not in sync, it still tries to compare each file. Most of the times
this comparison will fail, it goes for retry where will wait for checkpoint to
get streamed for 1 minute. This will happen for every file which differs causing
extremely long delay in reporting the failure.