Make DNS cache optional and remove duplicated code
The exported getDnsAddress() function was essentially a verbatbim copy
of getDnsCachedAddress(), but with the cache handling removed. Duplicate
code like this is always hard to keep in sync, and we had failed here as
getDnsAddress() didn't perform proper error handling and could lead to a
NULL pointer deref, where the cached version handled that. Refactor this
into a single function but retain the API to external callers. This also
renames the cache struct to match the naming convention of the file and
the codebase in general.

Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>