gprecoverseg: Do not throw exception in  `CleanSharedMem.execute()`.
The `CleanSharedMem.execute()` routine throws an exception if it cannot cleanup
shared memory.  However, there is a race condition between stopping the segments
and explicitly cleaning the shared memory.

This can happen when `CleanSharedMem.execute()` is able to read the
postmaster.pid file and find the shared memory to free, but the shared memory
has already been freed from previously stopping the segments with
`GpMirrorListToBuild.__ensurestopped()`.

Therefore, we are not throwing an exception in `CleanSharedMem.execute()` when
there is an error in cleaning up the shared memory when the shared memory does
not exist as it is most likely already freed from stopping the segements.

Co-authored-by: Shoaib Lari <slari@pivotal.io>
Co-authored-by: Kalen Krempely <kkrempely@pivotal.io>