pg_dump: try to omit expensive external partition queries
The query on the pg_partitions view to find external partitions is
expensive and currently runs for every partitioned table. Optimization
of the view itself has proven tricky -- many corner cases need to be
evaluated -- but it seems pointless to run the query if we know we don't
have any external partitions to begin with. This is especially true for
pg_upgrade, which currently disallows external partitions entirely.

Add a (much faster) preliminary query on pg_partition[_rule] to see if
the current table has any external partitions, and if not, skip the
expensive query entirely.