Simplify AO/CO diagnostic functions, by using 'regclass' arguments.
Instead of providing two varians of each function, one with table OID, and
another with table name as argument, use a single function with 'regclass'
argument. The implicit text to regclass cast takes care of looking up the
relation by name. So from a user's point of view, you can pass an OID or
a table name as argument, but the C code only sees the OID.

This is how, for example, the pg_relation_size() function works these days,
too. (See upstream commit 706a308806.)