Fix shell-quoting in COPY PROGRAM and EXECUTE external table.
The external program executed with COPY PROGRAM or EXECUTE-type external
table is passed a bunch of environment variables. They are passed by
adding them to the command line of the program being executed, with
"<var>=<value> && export VAR && ...". However, the quoting in the code
that builds that command line was broken. Fix it, and add a test.

Fixes github issue https://github.com/greenplum-db/gpdb/issues/5925