Revert "Fix shell-quoting in COPY PROGRAM and EXECUTE external table."
This reverts commit 7a9de421c8b13e61fc5f7c4a02070cfc74014452.

This caused regression failures like this:

--- /tmp/build/e18b2f02/gpdb_src/src/test/regress/expected/gpcopy.out	2018-10-11 07:42:24.231194166 +0000
+++ /tmp/build/e18b2f02/gpdb_src/src/test/regress/results/gpcopy.out	2018-10-11 07:42:24.287194166 +0000
@@ -1373,10 +1407,10 @@
 COPY foo FROM PROGRAM 'cat /tmp/gpcopyenvtest';
 COPY foo FROM PROGRAM 'echo database in COPY FROM: $GP_DATABASE';
 select * from foo;
-                        t
--------------------------------------------------
+                        t
+--------------------------------------------------
  data1
- database in COPY FROM: funny copy"db'withquotes
- database in COPY TO: funny copy"db'withquotes
+ database in COPY FROM: funny copy"db'with\quotes
+ database in COPY TO: funny copy"db'with\quotes
 (3 rows)

on the concourse pipeline, on SLES and Centos 6 and 7 systems. Looks like
the shell-quoting rules are different on those systems than on Debian /
Ubuntu. (More precisely, I guess the rules depend on the shell.)

The output with the backslash actually looks more correct to me, I wasn't
paying enough attention when I accepted the output without it, to the
expected output.