Fix indentation, to silence compiler warning.
After commit f8a78c255d, I started seeing this compiler warning:

check.c: In function ‘check_new_cluster’:
check.c:206:5: warning: this ‘if’ clause does not guard... [-Wmisleading-indentation]
     if (user_opts.segment_mode == DISPATCHER)
     ^~
check.c:210:2: note: ...this statement, but the latter is misleadingly indented as if it is guarded by the ‘if’
  check_for_prepared_transactions(&new_cluster);
  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Fix the indentation to use tabs instead of spaces, to silence that.

While we're at it, add an extra pair of braces, for clarity. I think
that's more clear, and Paul Guo also pointed at that in the original
review of the PR. This is in line of David Krieger's wish to keep the code
as close to upstream as possible.