Improve the INFO message on ORCA fallback.
If optimizer_trace_fallback is enabled, report the ORCA exception's message
to the user. It usually says something like "Feature not supported by the
Pivotal Query Optimizer: Non-default collation", which is quite helpful.

While we're at it, simplify setting the "had unexpected failure" flag,
by putting it in SOptContext. That allows removing one layer of catching and
re-throwing the exception.