Wait a little after setting GUC, to make test more robust.
The 'fts_recovery_in_progress' fails sometimes with a diff like this:

    --- /tmp/build/e18b2f02/gpdb_src/src/test/regress/expected/fts_recovery_in_progress.out	2018-10-12 01:35:16.895897976 +0000
    +++ /tmp/build/e18b2f02/gpdb_src/src/test/regress/results/fts_recovery_in_progress.out	2018-10-12 01:35:16.899897976 +0000
    @@ -37,7 +48,7 @@
     show gp_fts_probe_retries;
      gp_fts_probe_retries
     ----------------------
    - 2
    + 5
     (1 row)

The test sets the 'gp_fts_probe_retries' GUC with gpconfig, changing it
from 5 to 2, and sends SIGHUP, and then issues the above SHOW statement to
verify that the change took effect. But on a busy system, the SIGHUP and
the config change might not reach the backend quickly enough. Add a 5
second delay, to give it more time. On a very busy system, even that might
not of course be enough, but let's see how far we get with this, before we
start inventing a more complicated waiting mechanism.