Remove unnecessary debugging code, to reduce diff vs. upstream.
RESUME_INTERRUPTS() contains this same check, so this is redundant. The
message from RESUME_INTERRUPTS() is more generic, but a PANIC prints a
backtrace in the log, and possibly a core dump, and all the extra
information in these messages should be evident from the backtrace, too.
Besides, this isn't a very common failure mode,  I don't remember ever
hitting this PANIC. These extra checks were added a very long time ago,
maybe there was a specific problem that the developers were trying to hunt
down back then, but they don't seem useful anymore.

The upcoming 9.4 merge would further complicate these, or make them even
less useful. In 9.4, LWLocks are no longer identified by a simple integer,
so all we have here is a pointer to the LWLock struct, and parsing that
into a human-readable message would require more effort. Or we could just
print the pointer value, which is pretty useless.