Avoid deadlock between qp_functions_in_* tests.
The test functions modify the 'bar' table. If it's run in multiple
backends concurrently, they can deadlock. We've seen that causing failures
in the pipeline a few times.

To fix, have a separate copy of the 'foo' and 'bar' test tables for each
test that uses these functions.

Thanks to Ning Yu, Jialun Du and Melanie Plageman for hunting down the
cause of the deadlock.