Fix HINT on ALTER TABLE ... DISTRIBUTION KEY.
The DISTRIBUTED BY doesn't need to be a left-subset of the UNIQUE
indexes. Any subset will do. That has been true at least since e572af54e4,
although the check here has been the same even longer, so I suspect that
we allowed that for ALTER TABLE even before that.

Reviewed-by: Venkatesh Raghavan <vraghavan@pivotal.io>