pg_upgrade: don't run ANALYZE at all on segments
Before freezing tuples in the new target cluster, pg_upgrade performs an
ANALYZE to update pg_statistic. During segment upgrade, we have already
copied (master versions of the) AO auxiliary tables to the segment by
the time we reach this step, and the append-only analyze logic can't
deal with aoseg tuple counts that don't match the data on disk.
Therefore, we skip the ANALYZE step for segments.

Note that the incorrect master-only auxiliary table contents will be
overwritten later in the upgrade.