Remove references to reltoastidxid from gpcheckcat.
In the query in checkPGClass, AFAICS the reltoastrelid and reltoastidxid
fields were not used for anything, so just remove them from the query.
(To be honest, the whole query seems pretty pointless to me).

In the checkOwners() query, I also just removed the check for the toast
index. We have a regression test to check that the owner is set correctly
in ALTER TABLE commands, in the main test suite, in
'alter_distribution_policy' test, so it seems unlikely for that particular
problem to reappear. An index should always have the same owner as the
table itself, so perhaps adding a test for that more general case would be
useful, but I don't see much value in testing the toast index specifically.

pg_class.reltoastidxid field will be removed in PostgreSQL 9.4. Until
then, there's nothing wrong with these queries as such, but I wanted to
open a separate PR to get more eyeballs on the decisions.