Set correct numsegments for correlated SubPlan
Essentially we can regard SubPlan as an expression, it can be used to filter tuples. We have to broadcast the correlated relations which are in the SubPlan so that the Param can be used without crossing Slice.

In the previous codes, we broadcast the correlated relations according to the distribution of this relation, it may be wrong, we need to adjust it according to the position of SubPlan(or expression).

We add a flow to the context, then we can get the correct numsegments.