pipeline: Correctly map sqldump files
In PR https://github.com/greenplum-db/gpdb/pull/5996 during review, we
made the call to rename the resource used in dumping ICW.  This change
was not adaquately tested, and our initial search and replace in the
pipeline file name did not adequately take into account the dependancies
that exist between files used in Concourse.  This commit seeks to
correct those errors.

Authored-by: Jim Doty <jdoty@pivotal.io>