Refactor the way libpqwalreceiver is linked, to reduce diff vs. upstream.
In PostgreSQL, libpqwalreceiver is compiled into a separate shared library,
libpqwalreceiver.so, which is dynamically linked at runtime with dlopen().
It's done that way to avoid linking the backend with libpq. But GPDB has
its own, backend-specific, copy of libpq anyway, so we don't need to be
concerned about that.

Previously, libpqwalreceiver was replaced with a copy-pasted version,
gp_libpqwalreceiver.c, and the walrcv_* function pointers were replaced
with plain functions. If we were developing with no concern for merging,
that makes sense, since we don't need the indirection via function
pointers. But to reduce our diff vs. upstream, to make the merges easier,
it's better to have the function pointers, get rid of the separate
gp_libpqwalreceiver.c file, and only change the way libpqwalreceiver.c
is compiled, and how the "module", which is now just a regular .o file,
is loaded.